several bug fixes on ui,tommorow animations and group administration

+++ server/routes/api/socketEvents.js
@@ -3,47 +3,83 @@ const ObjectID = require("mongoose").Types.ObjectId;
 const fs = require("fs");
 const path = require("path");
 const xss = require("xss");
-
+const allowedTypes = [
+  "image/png","image/jpeg",
+  "image/jpg","image/svg","image/svg+xml",
+];
 async function onConnection(socket, io) {
   let profile = null;
   try {
     const pid = socket.handshake.session.passport.user._id;
     profile = await models.UsersModel.findById(pid);
+  
   } catch (err) {}
-  
-  socket.emit("auth", profile ? profile : null); 
-  
+
+  socket.emit("auth", profile ? profile : null);
+
   if (profile) {
-    const chatSet=new Set(profile.Chats.map(chat=>chat.toString()));
-    socket.use((stream,next)=>{console.log(1);next()})  
+    const chatSet = new Set(profile.Chats.map((chat) => chat.toString()));
+    chatSet.forEach((id)=>socket.join(id))
+
+    socket.use((stream, next) => {
+      console.log(1);
+      next();
+    });
     socket.on("logout", async () => {
       socket.emit("logout", async () => {});
     });
 
     async function SendMessage(message) {
-    
-      if(chatSet.has(message.cid)){
-      if(message.reply&&(!await models.MessagesModel.find({_id:new ObjectID(message.reply),cid:new ObjectID(message.cid)})))return;
-      console.log("ack")
-      const newMessage = new models.MessagesModel({
-        chat: new ObjectID(message.cid),
-        content: xss(message.content),
-        uid: profile._id,
-        reply_to:new ObjectID(message.reply)
-      });
-      newMessage
-        .save()
-        .catch((err) => console.log(err))
-        .then((data) => {
-          console.log(newMessage);
-          io.to(message.cid).emit(`messages`,{id:message.cid,replace:message.replace,data:[newMessage,]});
+
+      if (chatSet.has(message.cid)) {
+        if (message.reply)
+          message.reply=(await models.MessagesModel.find({
+          _id: new ObjectID(message.reply),
+          cid: new ObjectID(message.cid),
+        }))._id;
+
+        const newMessage = new models.MessagesModel({
+          chat: new ObjectID(message.cid),
+          content: xss(message.content),
+          uid: profile._id,
+          reply_to:(message.reply),
+          status:"✔"
         });
-    }}
+        newMessage
+          .save()
+          .catch((err) => console.log(err))
+          .then((data) => {
+            console.log("message",newMessage);
+            io.to(message.cid).emit(`messages`, {
+              id: message.cid,
+              replace: message.replace,
+              data: [newMessage],
+            });
+          });
+      }
+    }
+
+    socket.on("join",async function ([cid]){
+      const oid=new ObjectID(cid);
+      const chat=null
+      if(!chatSet.has(oid.toString())){
+       chat =await models.ChatsModel.findById(oid)
+       profile.Chats.push(chat._id);
+       chatSet.add(oid.toString());
+       chat.users.push(profile._id);
+       await profile.save();
+       await chat.save();
+       console.log("joined",chat);
+       socket.join(oid.toString());
+      }
+      if(chat)socket.emit("chat",{type:'join',chats:[chat]})   
+   
+    })
 
     socket.on("sendMessage", SendMessage);
 
     socket.on("search", async (stream) => {
-      //console.log(stream);
+      console.log(stream);
       term = new RegExp(xss(`${stream.query}`), "i");
       const model =
         stream.target == 1
@@ -65,44 +101,54 @@ async function onConnection(socket, io) {
             ],
           },
         },
-        {$project:{password:0,email:0}},
+        { $project: { password: 0, email: 0 } },
         { $sort: { username: 1 } },
         { $limit: 10 },
       ]);
       socket.emit("searchResults", { results: data, target: stream.target });
     });
-    socket.on("reportChat",(id)=>{
-
-    })
-    socket.on("deleteMessage",async ([mid,id,cid])=>{
-      let success=null;
-      if(chatSet.has(cid)){
-        success=await models.MessagesModel.findByIdAndDelete({_id:new ObjectID(id),chat:new ObjectID(cid)});
+    socket.on("reportChat", (id) => {});
+    
+    socket.on("deleteMessage", async ([mid, id, cid]) => {
+      let success = null;
+      console.log("delete")
+      if (chatSet.has(cid)) {
+        success = await models.MessagesModel.findByIdAndDelete({
+          _id: new ObjectID(id),
+          chat: new ObjectID(cid),
+        });
       }
-      if(!success)success={cid:cid,mid:mid,_id:id}
-      socket.emit(`deleteMessage`,[success._id,success.mid,success.chat])
-    })
-    socket.on("leaveChat",async({id,del})=>{
-      if(chatSet.has(id)){
-        console.log(id)
+      if (!success) success = { cid: cid, mid: mid, _id: id };
+      socket.emit(`deleteMessage`, [success._id, success.mid, success.chat]);
+    });
+    
+    socket.on("leaveChat", async ({ id, del }) => {
+      console.log("remove",id)
+      if (chatSet.has(id)) {
+        console.log(id);
         socket.leave(id);
-        profile.Chats=profile.Chats.filter((x)=>x!=id);
-        try{
-         const resp=await models.ChatsModel.findByIdAndUpdate(new ObjectID(id),{$pull:{users:profile._id}},{new:true});
-         console.log(resp)
-         (await models.ChatsModel.findOneAndDelete({_id:new ObjectID(id),type:'private'}));
-        }catch{}
+        profile.Chats = profile.Chats.filter((x) => x != id);
+        try {
+          const resp = await models.ChatsModel.findByIdAndUpdate(
+            new ObjectID(id),
+            { $pull: { users: profile._id } },
+            { new: true }
+          );
+          console.log(resp)(
+            await models.ChatsModel.findOneAndDelete({
+              _id: new ObjectID(id),
+              type: "private",
+            })
+          );
+        } catch {}
         await profile.save();
         chatSet.delete(id);
       }
-      socket.emit("leaveChat",[id,del])
-      
-    })
-    socket.on("muteChat",(id)=>{
-          
-    })
+      socket.emit("leaveChat", [id, del]);
+    });
+    socket.on("muteChat", (id) => {});
     socket.on("chats", async (stream) => {
-      const data =await models.ChatsModel.find({_id:{$in:[...chatSet]}});
+      const data = await models.ChatsModel.find({ _id: { $in: [...chatSet] } });
       socket.emit("chat", {
         type: stream.type,
         chats: data,
@@ -112,29 +158,34 @@ async function onConnection(socket, io) {
 
     socket.on("messages", async (stream) => {
       try {
-        const obj = { chat:new  ObjectID(stream.cid) };
-        if (stream.mid) obj.mid= (stream.gt)?{ $gt:stream.mid }:{ $lt:stream.mid };
-        console.log(obj)
-        const data = await models.MessagesModel.aggregate([
-          { $match: obj },                                
-          { $limit: 30 },                     
-          {
-            $lookup: {                         
-              from: 'messages',                
-              localField: 'reply_to',          
-              foreignField: '_id',             
-              as: 'replyToMessage'             
-            } 
-          },
-          { $unwind: {                         
-              path: "$replyToMessage",
-              preserveNullAndEmptyArrays: true 
-          }}
-        ]);
         
-        if (data.length){
-          io.to(stream.cid).emit(`messages`, {id:stream.cid,data:data});
-         console.log(data)
+         const obj = { chat: new ObjectID(stream.cid) };
+         if(!chatSet.has(stream.cid))obj.type='group'
+         if (stream.mid)
+          obj.mid = stream.gt ? { $gt: stream.mid } : { $lt: stream.mid };
+         
+         const data = await models.MessagesModel.aggregate([
+          { $match: obj },
+          { $limit: 30 },
+          {
+            $lookup: {
+              from: "messages",
+              localField: "reply_to",
+              foreignField: "_id",
+              as: "replyToMessage",
+            },
+          },
+          {
+            $unwind: {
+              path: "$replyToMessage",
+              preserveNullAndEmptyArrays: true,
+            },
+          },
+        ]);
+
+        if (data.length) {
+          io.to(stream.cid).emit(`messages`, { id: stream.cid, data: data });
+          console.log(data);
         }
       } catch (err) {
         console.log(err);
@@ -142,57 +193,47 @@ async function onConnection(socket, io) {
     });
 
     socket.on("updateProfile", async (stream) => {
-      if (stream.about || stream.name) {
         const changes = {};
-        changes[stream.about ? "about" : "name"] = stream.about || stream.name;
-        const object = await models.UsersModel.findByIdAndUpdate(profile._id, {
-          $set: changes,
-        }).then(() => {
-          profile[stream.about ? "about" : "name"] =
-            stream.about || stream.name;
-          socket.emit("profile", profile);
-        });
-      } else if (stream.img) {
-        const allowedTypes = [
-          "image/png",
-          "image/jpeg",
-          "image/jpg",
-          "image/svg",
-          "image/svg+xml",
-        ];
-        if (allowedTypes.includes(stream.img.type)) {
-          const buffer = Buffer.from(new Uint8Array(stream.file));
-          const pathto = path.join(`${profile._id}_${stream.img.name}`);
-          const content = {
-            src: pathto,
-            name: stream.img.name,
-            size: stream.img.size,
-            contentType: stream.img.type,
-            dimensions: stream.img.dimensions,
-          };
-          if (profile.img)
-            fs.unlink(
-              path.join(process.cwd(), "public", profile.img.src),
-              (err) => {}
-            );
-          fs.writeFile(
-            path.join(process.cwd(), "public", pathto),
-            buffer,
-            (err) => {
-              if (!err) {
-                models.UsersModel.findByIdAndUpdate(profile._id, {
-                  $set: { img: content },
-                }).then(() => {
-                  profile.img = content;
-                });
-              }
-            }
-          );
+        console.log("logged")
+        if (stream.about || stream.name) {
+                changes[stream.about?"about": "name"] = xss(stream.about || stream.name);
+                profile[stream.about?"about": "name"] = xss(stream.about || stream.name);
         }
-      }
-      chatSet.forEach((x)=>io.emit('profile',profile))
-      socket.emit("userProfile", profile);
+        if (stream.img&&allowedTypes.includes(stream.img.type)) {
+            const buffer = Buffer.from(new Uint8Array(stream.file));
+            const content = {
+              src: `${profile._id}_${stream.img.name}`,
+              name: stream.img.name,
+              size: stream.img.size,
+              contentType: stream.img.type,
+            };
+            if (profile.img)fs.unlink(path.join(process.cwd(), "public", profile.img.src));
+
+            fs.writeFile(
+              path.join(process.cwd(), "public", pathto),
+              buffer,
+              (err) => {
+                if (!err){profile.img=content;changes.img=content}
+              });
+            
+            }
+        console.log(changes)    
+        await profile.save();
+        changes._id=profile._id; 
+        socket.emit("profile", profile);
+       // chatSet.forEach((x) => io.to(x).emit("profile",changes));
     });
+
+    socket.on("updateChat", async (stream) =>{
+      const chat=await models.ChatsModel.findById(stream.cid) 
+      console.log("logged",chat.owner)
+     
+     
+      io.to(chat._id.toString()).emit("chat",{type:"chats",data:[chat]});
+    }
+    );
+
+
 
     socket.on("contacts", async (stream) => {
       socket.emit("contacts", profile.contacts);
@@ -203,23 +244,27 @@ async function onConnection(socket, io) {
     });
 
     socket.on("createChatPrivate", async (stream) => {
-
       let user = await models.UsersModel.findById(stream.cid);
-      if (user._id!=profile._id) {
-        const data = new models.ChatsModel({
-          users: [...new Set([profile._id, user._id])],
-          type: "private",
+      if (user._id != profile._id) {
+        let data = await models.ChatsModel.findOne({
+          users: [profile._id, user._id],
         });
-        const chat = await data.save();
-        chatSet.add(chat._id.toString());
-        profile.Chats.push(chat._id);
-        profile.contacts.push(user._id);
-        user.contacts.push(profile._id);
-        await profile.save();
-        await user.save();
-        chat.sender = stream.cid;
-        socket.join(chat._id.toString());
-        socket.emit(`private.${stream.cid}`, chat);
+        if (!data){
+           data = new models.ChatsModel({
+            users: [...new Set([profile._id, user._id])],
+            type: "private",
+          });
+          const chat = await data.save();
+          chatSet.add(chat._id.toString());
+          profile.Chats.push(chat._id);
+          profile.contacts.push(user._id);
+          user.contacts.push(profile._id);
+          await profile.save();
+          await user.save();
+          chat.sender = stream.cid;
+        }
+        socket.join(data._id.toString());
+        socket.emit(`private.${stream.cid}`,data);
         if (stream.content)
           SendMessage({
             content: stream.content,
